requirements.txt
together
Django>=4.2
python-decouple
supermemory